# openapi/router.yaml
openapi: 3.0.3
info:
  title: Hydra Router API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: OK
  /route:
    post:
      summary: Classify intent and choose reasoning path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Routing decision
          content:
            application/json:
              schema:
                type: object
                required: [routing_decision]
                properties:
                  routing_decision:
                    type: string
                    enum: [symbolic, llm, hybrid, create_teacher]
                  teacher_id:
                    type: string
                  normalized_input:
                    type: object
                    additionalProperties: true
                  notes:
                    type: string
  /normalize:
    post:
      summary: Turn free text into normalized inputs for a target teacher
      parameters:
        - in: query
          name: teacher_id
          schema: { type: string }
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query: { type: string }
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Normalized inputs
          content:
            application/json:
              schema:
                type: object
                properties:
                  normalized_input:
                    type: object
                    additionalProperties: true